+++ tests/plotting/test_general.py
@@ -6,7 +6,7 @@ import numpy as np
 import seaborn as sns
 import os
 import scanpy as sc
-import sctoolbox.plotting as pl
+import sctoolbox.plotting.general as pl
 
 from beartype.roar import BeartypeCallHintParamViolation