+++ tests/tools/__init__.py
@@ -1,5 +1,1 @@
-<<<<<<< HEAD
-"""Test init."""
-=======
 """Tools test init."""
->>>>>>> origin/dev