+++ tests/tools/test_celltype_annotation.py
@@ -4,7 +4,6 @@ import os
 import pytest
 import anndata as ad
 from sctoolbox.tools import celltype_annotation
-import sctoolbox.tools as anno
 import scanpy as sc
 
 # --------------------------- Fixtures ------------------------------ #
@@ -54,6 +53,6 @@ def test_add_cellxgene_annotation(adata_rna):
     """Test if 'cellxgene' column is added to adata.obs."""
 
     csv_f = os.path.join(os.path.dirname(__file__), '..', 'data', 'cellxgene_anno.csv')
-    anno.add_cellxgene_annotation(adata_rna, csv_f)
+    celltype_annotation.add_cellxgene_annotation(adata_rna, csv_f)
 
     assert "cellxgene_clusters" in adata_rna.obs.columns